Add a tool to translate problems to and from a linear git repository
This change introduces omegaup-translate-problem, which helps translate
git repositories for problems in the omegaUp quark format into git
repositories that are linear. This makes it a lot easier to rewrite
history since linear repositories are easier to work with. Additionally,
once history is rewritten, the packfiles can be regenerated so that only
the objects that are involved in the history are present.