Create an endpoint for renaming a repository (#7)
This change allows the frontend to rename a repository. This is needed
because when creating a problem, the metadata for it needs to be
committed into the database before sending out the request to gitserver,
and if this happens to fail, there is no way to undo this operation.

Instead, now the workflow will be create the repository with a random
name, and only if everything is ready to be committed, rename it to the
correct name, in a sort of two-phase commit.