Fix the health test (#30)
We were previously using the `sumas` problem for health tests. The
problem is that this is not always created. Furthermore, this makes it
possible for an exploit to mess with a production problem! So now we are
using the problem called `:testproblem` (with a colon at the beginning),
which is clearly not a valid problem name. This means that the very
first healthcheck request will create this problem, giving the extra
advantage of also testing for writability.